// src/api/rest/api/model/sm-source-dto.ts
/**
 * Shibametrics API
 * Shibametrics API Reference
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SMReportType } from './sm-report-type';


export interface SMSourceDto { 
    id?: number;
    name?: string | null;
    clientId?: number;
    type?: SMReportType;
    spreadSheetId?: string | null;
    rangeName?: string | null;
    rawData?: string | null;
    rawMetadata?: string | null;
}